chore(engine-v2): Add SSE subscription test
We were already testing this in the cli tests, so not a huge win but
I mostly run integration-tests so well.
