+++ Cargo.toml
@@ -71,6 +71,7 @@ async-graphql = "7.0.3"
 async-graphql-axum = "7.0.3"
 async-graphql-parser = "7.0.3"
 async-graphql-value = "7.0.3"
+async-sse = "5"
 async-trait = "0.1.80"
 axum = { version = "0.7.5", default-features = false }
 axum-server = { version = "0.6", default-features = false }
@@ -102,6 +103,7 @@ internment = { version = "0.8", features = ["serde", "arc"] }
 itertools = "0.13.0"
 jsonwebtoken = "9.3.0"
 governor = "0.6"
+multipart-stream = "0.1.2"
 num-traits = "0.2.18"
 once_cell = "1.19.0"
 openidconnect = "4.0.0-alpha.1"
