+++ engine/crates/integration-tests/Cargo.toml
@@ -12,7 +12,9 @@ async-graphql-parser.workspace = true
 async-graphql.workspace = true
 async-once-cell = "0.5.3"
 async-runtime.workspace = true
+async-sse.workspace = true
 async-trait.workspace = true
+bytes.workspace = true
 crossbeam-queue = "0.3"
 cynic.workspace = true
 cynic-introspection.workspace = true
@@ -32,7 +34,7 @@ headers.workspace = true
 http.workspace = true
 indoc = "2.0.5"
 insta.workspace = true
-multipart-stream = "0.1.2"
+multipart-stream.workspace = true
 names = "0.14.1-dev"
 openidconnect.workspace = true
 reqwest.workspace = true
