+++ engine/crates/integration-tests/tests/federation/subscriptions/mod.rs
@@ -1,0 +1,2 @@
+mod multipart;
+mod sse;
